requirements.txt
lightstreamer-client-lib==2.2.2